CHG: move logging setup to function
