+++ syncopy/__init__.py
@@ -104,68 +104,6 @@ else:
     else:
         __storage__ = os.path.join(os.path.expanduser("~"), ".spy", "tmp_storage")
 
-# Setup logging.
-if os.environ.get("SPYLOGDIR"):
-    __logdir__ = os.path.abspath(os.path.expanduser(os.environ["SPYLOGDIR"]))
-else:
-    if os.path.exists(csHome):
-        __logdir__ = os.path.join(csHome, ".spy", "logs")
-    else:
-        __logdir__ = os.path.join(os.path.expanduser("~"), ".spy", "logs")
-
-if not os.path.exists(__logdir__):
-    os.makedirs(__logdir__, exist_ok=True)
-
-loglevel = os.getenv("SPYLOGLEVEL", "WARNING")
-numeric_level = getattr(logging, loglevel.upper(), None)
-if not isinstance(numeric_level, int):  # An invalid string was set as the env variable, default to WARNING.
-    warnings.warn("Invalid log level set in environment variable 'SPYLOGLEVEL', ignoring and using WARNING instead. Hint: Set one of 'DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'.")
-    loglevel = "WARNING"
-
-# The logger for local/sequential stuff -- goes to terminal and to a file.
-spy_logger = logging.getLogger('syncopy')
-fmt = logging.Formatter('%(asctime)s - %(levelname)s: %(message)s')
-sh = logging.StreamHandler(sys.stdout)
-sh.setFormatter(fmt)
-spy_logger.addHandler(sh)
-
-logfile = os.path.join(__logdir__, f'syncopy.log')
-fh = logging.FileHandler(logfile)  # The default mode is 'append'.
-fh.setFormatter(fmt)
-spy_logger.addHandler(fh)
-
-
-spy_logger.setLevel(loglevel)
-spy_logger.debug(f"Starting Syncopy session at {datetime.datetime.now().astimezone().isoformat()}.")
-spy_logger.info(f"Syncopy log level set to: {loglevel}.")
-
-# Log to per-host files in parallel code by default.
-# Note that this setup handles only the logger of the current host.
-parloglevel = os.getenv("SPYPARLOGLEVEL", loglevel)
-numeric_level = getattr(logging, parloglevel.upper(), None)
-if not isinstance(numeric_level, int):  # An invalid string was set as the env variable, use default.
-    warnings.warn("Invalid log level set in environment variable 'SPYPARLOGLEVEL', ignoring and using WARNING instead. Hint: Set one of 'DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'.")
-    parloglevel = "WARNING"
-host = platform.node()
-parallel_logger_name = "syncopy_" + host
-spy_parallel_logger = logging.getLogger(parallel_logger_name)
-
-class HostnameFilter(logging.Filter):
-    hostname = platform.node()
-
-    def filter(self, record):
-        record.hostname = HostnameFilter.hostname
-        return True
-
-logfile_par = os.path.join(__logdir__, f'syncopy_{host}.log')
-fhp = logging.FileHandler(logfile_par)  # The default mode is 'append'.
-fhp.addFilter(HostnameFilter())
-spy_parallel_logger.setLevel(parloglevel)
-fmt_with_hostname = logging.Formatter('%(asctime)s - %(levelname)s - %(hostname)s: %(message)s')
-fhp.setFormatter(fmt_with_hostname)
-spy_parallel_logger.addHandler(fhp)
-spy_parallel_logger.info(f"Syncopy parallel logger '{parallel_logger_name}' setup to log to file '{logfile_par}' at level {loglevel}.")
-
 
 # Set upper bound for temp directory size (in GB)
 __storagelimit__ = 10
@@ -194,6 +132,10 @@ from .nwanalysis import *
 from .statistics import *
 from .plotting import *
 from .preproc import *
+
+from .shared.log import setup_logging
+setup_logging()
+
 
 # Register session
 __session__ = datatype.util.SessionLogger()
