+++ syncopy/shared/log.py
@@ -5,14 +5,86 @@
 # Note: The logging setup is done in the top-level `__init.py__` file.
 
 import os
+import sys
 import logging
 import socket
 import syncopy
 import warnings
+import datetime
+import platform
+import getpass
+import syncopy
 
 
 loggername = "syncopy"  # Since this is a library, we should not use the root logger (see Python logging docs).
 loglevels = ['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL']
+
+def setup_logging():
+    print("setting up logging")
+    # Setup logging.
+    csHome = "/cs/home/{}".format(getpass.getuser())
+    if os.environ.get("SPYLOGDIR"):
+        syncopy.__logdir__ = os.path.abspath(os.path.expanduser(os.environ["SPYLOGDIR"]))
+    else:
+        if os.path.exists(csHome):
+            syncopy.__logdir__ = os.path.join(csHome, ".spy", "logs")
+        else:
+            syncopy.__logdir__ = os.path.join(os.path.expanduser("~"), ".spy", "logs")
+
+    if not os.path.exists(syncopy.__logdir__):
+        os.makedirs(syncopy.__logdir__, exist_ok=True)
+
+    loglevel = os.getenv("SPYLOGLEVEL", "WARNING")
+    numeric_level = getattr(logging, loglevel.upper(), None)
+    if not isinstance(numeric_level, int):  # An invalid string was set as the env variable, default to WARNING.
+        warnings.warn("Invalid log level set in environment variable 'SPYLOGLEVEL', ignoring and using WARNING instead. Hint: Set one of 'DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'.")
+        loglevel = "WARNING"
+
+    # The logger for local/sequential stuff -- goes to terminal and to a file.
+    spy_logger = logging.getLogger('syncopy')
+    fmt = logging.Formatter('%(asctime)s - %(levelname)s: %(message)s')
+    sh = logging.StreamHandler(sys.stdout)
+    sh.setFormatter(fmt)
+    spy_logger.addHandler(sh)
+
+    logfile = os.path.join(syncopy.__logdir__, f'syncopy.log')
+    fh = logging.FileHandler(logfile)  # The default mode is 'append'.
+    fh.setFormatter(fmt)
+    spy_logger.addHandler(fh)
+
+
+    spy_logger.setLevel(loglevel)
+    spy_logger.debug(f"Starting Syncopy session at {datetime.datetime.now().astimezone().isoformat()}.")
+    spy_logger.info(f"Syncopy log level set to: {loglevel}.")
+
+    # Log to per-host files in parallel code by default.
+    # Note that this setup handles only the logger of the current host.
+    parloglevel = os.getenv("SPYPARLOGLEVEL", loglevel)
+    numeric_level = getattr(logging, parloglevel.upper(), None)
+    if not isinstance(numeric_level, int):  # An invalid string was set as the env variable, use default.
+        warnings.warn("Invalid log level set in environment variable 'SPYPARLOGLEVEL', ignoring and using WARNING instead. Hint: Set one of 'DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'.")
+        parloglevel = "WARNING"
+    host = platform.node()
+    parallel_logger_name = "syncopy_" + host
+    spy_parallel_logger = logging.getLogger(parallel_logger_name)
+
+    class HostnameFilter(logging.Filter):
+        hostname = platform.node()
+
+        def filter(self, record):
+            record.hostname = HostnameFilter.hostname
+            return True
+
+    logfile_par = os.path.join(syncopy.__logdir__, f'syncopy_{host}.log')
+    fhp = logging.FileHandler(logfile_par)  # The default mode is 'append'.
+    fhp.addFilter(HostnameFilter())
+    spy_parallel_logger.setLevel(parloglevel)
+    fmt_with_hostname = logging.Formatter('%(asctime)s - %(levelname)s - %(hostname)s: %(message)s')
+    fhp.setFormatter(fmt_with_hostname)
+    spy_parallel_logger.addHandler(fhp)
+    spy_parallel_logger.info(f"Syncopy parallel logger '{parallel_logger_name}' setup to log to file '{logfile_par}' at level {loglevel}.")
+
+
 
 def get_logger():
     """Get the syncopy root logger.
